Fix casts
OCaml ints are really longs, so use the correct type cast.

+++ lib/nativeint_conv.c
@@ -30,56 +30,56 @@ CAMLprim value
 nativeint_of_int8(value v)
 {
   CAMLparam1(v);
-  CAMLreturn (caml_copy_nativeint((int)Int8_val(v)));
+  CAMLreturn (caml_copy_nativeint((long)Int8_val(v)));
 }
 
 CAMLprim value
 nativeint_of_int16(value v)
 {
   CAMLparam1(v);
-  CAMLreturn (caml_copy_nativeint((int)Int16_val(v)));
+  CAMLreturn (caml_copy_nativeint((long)Int16_val(v)));
 }
 
 CAMLprim value
 nativeint_of_int24(value v)
 {
   CAMLparam1(v);
-  CAMLreturn (caml_copy_nativeint((int)Int24_val(v)));
+  CAMLreturn (caml_copy_nativeint((long)Int24_val(v)));
 }
 
 CAMLprim value
 nativeint_of_int32(value v)
 {
   CAMLparam1(v);
-  CAMLreturn (caml_copy_nativeint((int)Int32_val(v)));
+  CAMLreturn (caml_copy_nativeint((long)Int32_val(v)));
 }
 
 CAMLprim value
 nativeint_of_int40(value v)
 {
   CAMLparam1(v);
-  CAMLreturn (caml_copy_nativeint((int)Int40_val(v)));
+  CAMLreturn (caml_copy_nativeint((long)Int40_val(v)));
 }
 
 CAMLprim value
 nativeint_of_int48(value v)
 {
   CAMLparam1(v);
-  CAMLreturn (caml_copy_nativeint((int)Int48_val(v)));
+  CAMLreturn (caml_copy_nativeint((long)Int48_val(v)));
 }
 
 CAMLprim value
 nativeint_of_int56(value v)
 {
   CAMLparam1(v);
-  CAMLreturn (caml_copy_nativeint((int)Int56_val(v)));
+  CAMLreturn (caml_copy_nativeint((long)Int56_val(v)));
 }
 
 CAMLprim value
 nativeint_of_int64(value v)
 {
   CAMLparam1(v);
-  CAMLreturn (caml_copy_nativeint((int)Int64_val(v)));
+  CAMLreturn (caml_copy_nativeint((long)Int64_val(v)));
 }
 
 CAMLprim value
@@ -87,7 +87,7 @@ nativeint_of_int128(value v)
 {
   CAMLparam1(v);
 #ifdef HAVE_INT128
-  CAMLreturn (caml_copy_nativeint((int)Int128_val(v)));
+  CAMLreturn (caml_copy_nativeint((long)Int128_val(v)));
 #else
   failwith("unimplemented");
   CAMLreturn(Val_unit);
@@ -98,56 +98,56 @@ CAMLprim value
 nativeint_of_uint8(value v)
 {
   CAMLparam1(v);
-  CAMLreturn (caml_copy_nativeint((int)Uint8_val(v)));
+  CAMLreturn (caml_copy_nativeint((long)Uint8_val(v)));
 }
 
 CAMLprim value
 nativeint_of_uint16(value v)
 {
   CAMLparam1(v);
-  CAMLreturn (caml_copy_nativeint((int)Uint16_val(v)));
+  CAMLreturn (caml_copy_nativeint((long)Uint16_val(v)));
 }
 
 CAMLprim value
 nativeint_of_uint24(value v)
 {
   CAMLparam1(v);
-  CAMLreturn (caml_copy_nativeint((int)Uint24_val(v)));
+  CAMLreturn (caml_copy_nativeint((long)Uint24_val(v)));
 }
 
 CAMLprim value
 nativeint_of_uint32(value v)
 {
   CAMLparam1(v);
-  CAMLreturn (caml_copy_nativeint((int)Uint32_val(v)));
+  CAMLreturn (caml_copy_nativeint((long)Uint32_val(v)));
 }
 
 CAMLprim value
 nativeint_of_uint40(value v)
 {
   CAMLparam1(v);
-  CAMLreturn (caml_copy_nativeint((int)Uint40_val(v)));
+  CAMLreturn (caml_copy_nativeint((long)Uint40_val(v)));
 }
 
 CAMLprim value
 nativeint_of_uint48(value v)
 {
   CAMLparam1(v);
-  CAMLreturn (caml_copy_nativeint((int)Uint48_val(v)));
+  CAMLreturn (caml_copy_nativeint((long)Uint48_val(v)));
 }
 
 CAMLprim value
 nativeint_of_uint56(value v)
 {
   CAMLparam1(v);
-  CAMLreturn (caml_copy_nativeint((int)Uint56_val(v)));
+  CAMLreturn (caml_copy_nativeint((long)Uint56_val(v)));
 }
 
 CAMLprim value
 nativeint_of_uint64(value v)
 {
   CAMLparam1(v);
-  CAMLreturn (caml_copy_nativeint((int)Uint64_val(v)));
+  CAMLreturn (caml_copy_nativeint((long)Uint64_val(v)));
 }
 
 CAMLprim value
@@ -155,7 +155,7 @@ nativeint_of_uint128(value v)
 {
   CAMLparam1(v);
 #ifdef HAVE_UINT128
-  CAMLreturn (caml_copy_nativeint((int)Uint128_val(v)));
+  CAMLreturn (caml_copy_nativeint((long)Uint128_val(v)));
 #else
   failwith("unimplemented");
   CAMLreturn(Val_unit);
